feat: subscription system, notification bell, shifts payment removal
- Add WorkerSubscription model + WorkerPlan enum to Prisma schema
- Add SubscriptionService with plan limits for workers (free/premium) and employers (free/basic/pro)
- Add /api/v1/subscriptions routes (worker/me, worker/checkout, employer/me, employer/checkout)
- Webhook handles subscription payments, grants plan and sends notification
- Worker application limit enforced (5/month on free, unlimited on premium)
- Employer vacancy limit enforced (3 active on free, 15 on basic, unlimited on pro)
- Admin can manually grant subscriptions via PATCH /admin/users/:userId/subscription
- Add GET /admin/subscriptions to list all active paid subscriptions
- Add worker subscription page with Free/Premium plan cards and checkout
- Add employer subscription page with Старт/Бизнес/Про plan cards and checkout
- Add NotificationBell component in DashboardTopBar for both cabinets
- Update sidebar links: replace earnings/payments with subscription page
- Remove payment processing UI from employer shifts (tab, column, button)
- Remove payment status display from worker shifts
- Fix employer invitations: show all statuses, not only invited
- Fix canReview: no longer requires hasPaid

+++ packages/shared/src/schemas.ts
@@ -185,7 +185,7 @@ export const workerProfileUpdateSchema = z.object({
   lastName: z.string().min(2).max(50).optional(),
   age: z.number().int().min(18).max(99).optional(),
   bio: z.string().max(2000).optional(),
-  experienceYears: z.number().int().min(0).max(50).optional(),
+  experienceYears: z.number().min(0).max(50).optional(),
   hasMedicalBook: z.boolean().optional(),
   desiredRate: z.number().positive().optional(),
   rateType: z.enum(['hourly', 'per_shift', 'fixed', 'daily', 'weekly', 'after_event']).optional(),
